test: run browser-capable coverage in Chromium
